# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CRM",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.0.120:8000",
    "paths": {
...
        "/api/v1/chuc-vu": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Filter role based on provided filters",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "role"
                ],
                "summary": "Filter Role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filters in JSON format",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc/desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing role entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "Update Role",
                "parameters": [
                    {
                        "description": "Updated Role data",
                        "name": "Chuc_vu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Chuc_vu_update"
                        }
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new role entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "Create Role",
                "parameters": [
                    {
                        "description": "Role data",
                        "name": "Discount_Type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Chuc_vu_create"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/chuc-vu/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an existing role entry",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "Delete Role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID to be deleted",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/cong-no-khach-hang": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Filter customer debt based on provided filters",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "debt"
                ],
                "summary": "Filter Customer Debt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filters in JSON format",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc/desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/cong-no-nha-phan-phoi": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Filter provider debt based on provided filters",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "debt"
                ],
                "summary": "Filter Provider Debt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filters in JSON format",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc/desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/dang-nhap": {
            "post": {
                "description": "Login API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login data include username, password",
                        "name": "Login_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Dang_nhap"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Dang_nhap"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/thong-tin-nhan-vien": {
            "get": {
                "description": "Get Me API by sending JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get Me",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Get_me"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.Chuc_vu_create": {
            "type": "object",
            "required": [
                "ten"
            ],
            "properties": {
                "ten": {
                    "type": "string"
                }
            }
        },
        "requests.Chuc_vu_update": {
            "type": "object",
            "required": [
                "id",
                "ten"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "ten": {
                    "type": "string"
                }
            }
        },
        "requests.Dang_nhap": {
            "type": "object",
            "required": [
                "mat_khau",
                "ten_dang_nhap"
            ],
            "properties": {
                "mat_khau": {
                    "type": "string"
                },
                "ten_dang_nhap": {
                    "type": "string"
                }
            }
        },
...
        "responses.Dang_nhap": {
            "type": "object",
            "properties": {
                "ds_quyen": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "responses.Get_me": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "UpdateAt": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "chuc_vu": {
                    "type": "string"
                },
                "chuc_vu_id": {
                    "type": "string"
                },
                "dia_chi": {
                    "type": "string"
                },
                "dien_thoai": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "ho_ten": {
                    "type": "string"
                },
                "quyen": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ten_dang_nhap": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}